+++ zenml/datasources/base_datasource_test.py
@@ -17,9 +17,9 @@ import os
 import zenml
 import random
 import pandas as pd
-from zenml.core.datasources import BaseDatasource
-from zenml.core.pipelines import BasePipeline
-from zenml.core.standards import standard_keys as keys
+from zenml.datasources import BaseDatasource
+from zenml.pipelines import BasePipeline
+from zenml.standards import standard_keys as keys
 from zenml.utils import yaml_utils, exceptions
 from pathlib import Path
